requirements.txt
paho-mqtt==1.2
pi_switch==0.4.0
